Fixed a bug within handleArg() method, changed the value from double to int type for when dealing with the PUT command SET_WATER_LEVEL

+++ systemh-backend/src/edu/hawaii/ihale/backend/restserver/resource/PutResource.java
@@ -96,7 +96,7 @@ public class PutResource extends ServerResource {
       commandArg = Double.parseDouble(arg);
       break;
     case SET_WATER_LEVEL:
-      commandArg = Double.parseDouble(arg);
+      commandArg = Integer.parseInt(arg);
       break;
     // Lighting
     case SET_LIGHTING_LEVEL:
